+++ CHANGELOG.md
@@ -298,6 +298,11 @@ IMPROVEMENTS:
  * driver/rkt: Support `no_overlay` [[GH-2702](https://github.com/hashicorp/nomad/issues/2702)]
  * driver/rkt: Support `insecure_options` list [[GH-2695](https://github.com/hashicorp/nomad/issues/2695)]
  * server: Allow tuning of node heartbeat TTLs [[GH-2859](https://github.com/hashicorp/nomad/issues/2859)]
+   [GH-2535]
+ * driver/docker: Allow using soft memory limit instead of hard one [GH-2771]
+ * driver/rkt: Support `no_overlay` [GH-2702]
+ * driver/rkt: Support `insecure_options` list [GH-2695]
+ * server: Allow tuning of node heartbeat TTLs [GH-2859]
  * server/networking: Shrink dynamic port range to not overlap with majority of
    operating system's ephemeral port ranges to avoid port conflicts [[GH-2856](https://github.com/hashicorp/nomad/issues/2856)]